+++ data-plane/core/config/src/tls/root_store_builder.rs
@@ -126,8 +126,7 @@ impl RootStoreBuilder {
 
         if !spiffe_cfg.trust_domains.is_empty() {
             for domain in &spiffe_cfg.trust_domains {
-                let bundle = spire_identity_manager
-                    .get_x509_bundle_for_trust_domain(domain)?;
+                let bundle = spire_identity_manager.get_x509_bundle_for_trust_domain(domain)?;
 
                 for cert in bundle.authorities() {
                     let der_cert = CertificateDer::from(cert.as_ref().to_vec());