+++ data-plane/core/config/CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.2](https://github.com/agntcy/slim/compare/slim-config-v0.8.1...slim-config-v0.8.2) - 2026-03-31
+
+### Other
+
+- update Cargo.lock dependencies
+
+## [0.8.1](https://github.com/agntcy/slim/compare/slim-config-v0.8.0...slim-config-v0.8.1) - 2026-03-30
+
+### Other
+
+- update Cargo.lock dependencies
+
+## [0.8.0](https://github.com/agntcy/slim/compare/slim-config-v0.7.0...slim-config-v0.8.0) - 2026-03-26
+
+### Added
+
+- MLS identity key integration and security dependency upgrades ([#1394](https://github.com/agntcy/slim/pull/1394))
+- add link negotiation protocol between SLIM nodes ([#1353](https://github.com/agntcy/slim/pull/1353))
+
 ## [0.7.0](https://github.com/agntcy/slim/compare/slim-config-v0.6.3...slim-config-v0.7.0) - 2026-03-20
 
 ### Added