+++ data-plane/core/tracing/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.8](https://github.com/agntcy/slim/compare/slim-tracing-v0.3.7...slim-tracing-v0.3.8) - 2026-03-31
+
+### Other
+
+- updated the following local packages: agntcy-slim-version, agntcy-slim-config
+
+## [0.3.7](https://github.com/agntcy/slim/compare/slim-tracing-v0.3.6...slim-tracing-v0.3.7) - 2026-03-30
+
+### Other
+
+- updated the following local packages: agntcy-slim-version, agntcy-slim-config
+
+## [0.3.6](https://github.com/agntcy/slim/compare/slim-tracing-v0.3.5...slim-tracing-v0.3.6) - 2026-03-26
+
+### Other
+
+- updated the following local packages: agntcy-slim-version, agntcy-slim-config
+
 ## [0.3.5](https://github.com/agntcy/slim/compare/slim-tracing-v0.3.4...slim-tracing-v0.3.5) - 2026-03-20
 
 ### Added