+++ data-plane/slimrpc-compiler/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0](https://github.com/agntcy/slim/compare/protoc-slimrpc-plugin-v1.2.0...protoc-slimrpc-plugin-v1.3.0) - 2026-03-31
+
+### Added
+
+- slimrcp multicast examples ([#1346](https://github.com/agntcy/slim/pull/1346))
+- Fix dotnet bindings ([#1403](https://github.com/agntcy/slim/pull/1403))
+- Add slimrpc support for .NET (C#) ([#1345](https://github.com/agntcy/slim/pull/1345))
+
 ## [1.2.0](https://github.com/agntcy/slim/compare/protoc-slimrpc-plugin-v1.1.0...protoc-slimrpc-plugin-v1.2.0) - 2026-03-20
 
 ### Added